+++ platform.pde
@@ -2,9 +2,9 @@ public class platform {
   
    private float x, y, w, h;
    
-   public platform(int x, int y, int w, int h) {
+   public platform(float x, float y, float w, float h) {
     this.x = x;
-    this.y = y;
+     this.y = y;
     this.w = w;
     this.h = h;
    }
